+++ tests/unit/build-test.ts
@@ -7,28 +7,28 @@ import fileB from 'dummy/b';
 import { description as fromAts } from 'in-repo-a/test-support/from-ats';
 import { description as fromTs } from 'dummy/tests/from-ts';
 
-module('Unit | Build', function() {
-  test("in-repo addons' addon trees wind up in the right place", function(assert) {
+module('Unit | Build', function () {
+  test("in-repo addons' addon trees wind up in the right place", function (assert) {
     assert.equal(addonFileA, 'in-repo-a/test-file');
     assert.equal(addonFileB, 'in-repo-b/test-file');
   });
 
-  test("in-repo addons' app trees wind up in the right place", function(assert) {
+  test("in-repo addons' app trees wind up in the right place", function (assert) {
     assert.equal(fileA, 'dummy/a');
     assert.equal(fileB, 'dummy/b');
   });
 
-  test("addon's addon-test-support files end up in <addon-name>/test-support/*", function(assert) {
+  test("addon's addon-test-support files end up in <addon-name>/test-support/*", function (assert) {
     assert.ok(fromAts);
     assert.equal(fromAts, 'From addon-test-support');
   });
 
-  test("addon's test-support files end up in dummy/tests/*", function(assert) {
+  test("addon's test-support files end up in dummy/tests/*", function (assert) {
     assert.ok(fromTs);
     assert.equal(fromTs, 'From test-support');
   });
 
-  test('optional chaining and nullish coalescing are transpiled correctly', function(assert) {
+  test('optional chaining and nullish coalescing are transpiled correctly', function (assert) {
     let value = { a: 'hello' } as { a?: string; b?: string };
     assert.equal(value?.a, 'hello');
     assert.equal(value?.b, undefined);
@@ -36,7 +36,7 @@ module('Unit | Build', function() {
     assert.equal(value?.b ?? 'ok', 'ok');
   });
 
-  test('class field declarations work', function(assert) {
+  test('class field declarations work', function (assert) {
     class MyClass {
       declare foo: string;
     }