+++ ts/tests/helpers/skeleton-app.ts
@@ -1,12 +1,10 @@
 import fs from 'fs-extra';
 import path from 'path';
-import tmp from 'tmp';
 import execa from 'execa';
 import { EventEmitter } from 'events';
 import got from 'got';
 import debugLib from 'debug';
-
-tmp.setGracefulCleanup();
+import rimraf from 'rimraf';
 
 const debug = debugLib('skeleton-app');
 
@@ -18,16 +16,13 @@ const getEmberPort = (() => {
 export default class SkeletonApp {
   port = getEmberPort();
   watched: WatchedBuild | null = null;
-  tmpDir = tmp.dirSync({
-    tries: 10,
-    unsafeCleanup: true,
-    dir: process.cwd(),
-    template: 'test-skeleton-app-XXXXXX',
-  });
-  root = this.tmpDir.name;
+  cleanupTempDir = () => rimraf(this.root, (error) => console.error(error));
+  root = path.join(process.cwd(), `test-skeleton-app-${Math.random().toString(36).slice(2)}`);
 
   constructor() {
+    fs.mkdirpSync(this.root);
     fs.copySync(`${__dirname}/../../../test-fixtures/skeleton-app`, this.root);
+    process.on('beforeExit', this.cleanupTempDir);
   }
 
   build() {
@@ -68,7 +63,9 @@ export default class SkeletonApp {
     if (this.watched) {
       this.watched.kill();
     }
-    this.tmpDir.removeCallback();
+
+    this.cleanupTempDir();
+    process.off('beforeExit', this.cleanupTempDir);
   }
 
   _ember(args: string[]) {
@@ -80,7 +77,7 @@ export default class SkeletonApp {
 class WatchedBuild extends EventEmitter {
   constructor(protected ember: execa.ExecaChildProcess, protected port: number) {
     super();
-    this.ember.stdout.on('data', data => {
+    this.ember.stdout?.on('data', (data) => {
       let output = data.toString();
       if (output.includes('Build successful')) {
         this.emit('did-rebuild');
@@ -89,11 +86,11 @@ class WatchedBuild extends EventEmitter {
       debug(output);
     });
 
-    this.ember.stderr.on('data', data => {
+    this.ember.stderr?.on('data', (data) => {
       debug(data.toString());
     });
 
-    this.ember.catch(error => {
+    this.ember.catch((error) => {
       this.emit('did-error', error);
     });
   }
@@ -103,19 +100,19 @@ class WatchedBuild extends EventEmitter {
   }
 
   waitForOutput(target: string) {
-    return new Promise(resolve => {
+    return new Promise((resolve) => {
       let output = '';
       let listener = (data: string | Buffer) => {
         output += data.toString();
         if (output.includes(target)) {
-          this.ember.stdout.removeListener('data', listener);
-          this.ember.stderr.removeListener('data', listener);
+          this.ember.stdout?.removeListener('data', listener);
+          this.ember.stderr?.removeListener('data', listener);
           resolve(output);
         }
       };
 
-      this.ember.stdout.on('data', listener);
-      this.ember.stderr.on('data', listener);
+      this.ember.stdout?.on('data', listener);
+      this.ember.stderr?.on('data', listener);
     });
   }